computer map added

+++ server/controllers/adminController.js
@@ -1,7 +1,6 @@
 import User from "../models/User.js";
 import bcrypt from "bcrypt";
 
-// 1. Lấy danh sách user
 export const getAllUsers = async (req, res) => {
   try {
     const users = await User.findAll({
@@ -14,20 +13,22 @@ export const getAllUsers = async (req, res) => {
   }
 };
 
-// 2. Tạo tài khoản mới (Create)
 export const createUser = async (req, res) => {
   const { user_name, password, role_id } = req.body;
   try {
-    // Kiểm tra trùng tên
     const exists = await User.findOne({ where: { user_name } });
     if (exists)
       return res.status(400).json({ message: "Tên đăng nhập đã tồn tại." });
+
+    if (parseInt(role_id) === 1) {
+      return res.status(400).json({ message: "Không thể tạo thêm Admin." });
+    }
 
     const hashedPassword = await bcrypt.hash(password, 10);
     const newUser = await User.create({
       user_name,
       password: hashedPassword,
-      role_id: role_id || 3, // Mặc định là User
+      role_id: role_id || 3,
       balance: 0,
     });
 
@@ -35,14 +36,17 @@ export const createUser = async (req, res) => {
       .status(201)
       .json({ message: "Tạo tài khoản thành công!", user: newUser });
   } catch (error) {
-    console.error(error); // Log lỗi ra terminal để dễ debug
     res.status(500).json({ message: "Lỗi tạo tài khoản." });
   }
 };
 
-// 3. Xóa tài khoản (Delete)
 export const deleteUser = async (req, res) => {
   const { id } = req.params;
+
+  if (parseInt(id) === 1) {
+    return res.status(403).json({ message: "Không thể xóa Admin gốc!" });
+  }
+
   try {
     await User.destroy({ where: { user_id: id } });
     res.json({ message: "Đã xóa tài khoản." });
@@ -51,7 +55,6 @@ export const deleteUser = async (req, res) => {
   }
 };
 
-// 4. Nạp tiền (Top-up Balance)
 export const topUpBalance = async (req, res) => {
   const { id } = req.params;
   const { amount } = req.body;
@@ -66,5 +69,36 @@ export const topUpBalance = async (req, res) => {
     res.json({ message: `Đã nạp ${amount} VNĐ. Số dư mới: ${user.balance}` });
   } catch (error) {
     res.status(500).json({ message: "Lỗi nạp tiền." });
+  }
+};
+
+export const changeUserRole = async (req, res) => {
+  const { id } = req.params;
+  const { role_id } = req.body;
+
+  if (parseInt(id) === 1) {
+    return res
+      .status(403)
+      .json({ message: "Không thể thay đổi quyền của Admin gốc." });
+  }
+
+  if (parseInt(role_id) === 1) {
+    return res
+      .status(400)
+      .json({ message: "Chỉ có duy nhất 1 Admin. Không thể cấp quyền này." });
+  }
+
+  try {
+    const user = await User.findByPk(id);
+    if (!user) {
+      return res.status(404).json({ message: "Người dùng không tồn tại." });
+    }
+
+    user.role_id = parseInt(role_id);
+    await user.save();
+
+    res.json({ message: `Đã cập nhật quyền thành công!` });
+  } catch (error) {
+    res.status(500).json({ message: "Lỗi khi đổi quyền." });
   }
 };
